Updated preprocessing functions

+++ unmicst_s3_non_snr/tumor_out_patient_hyper/combine_datasets.py
@@ -11,12 +11,13 @@ if __name__ == '__main__':
     target = args.target  # The target biopsy which should be excluded
     directory = args.dir
 
-    train = []
-    for root, dirs, files in os.walk(directory):
-        for name in files:
-            if Path(name).suffix == ".csv" and args.target not in name and "excluded" not in name:
-                print("Loading {}".format(name))
-                train.append(pd.read_csv(os.path.join(root, name), sep="\t", header=0))
+    if not Path(f"{args.output_dir}/{target}_excluded_dataset.csv").exists():
+        train = []
+        for root, dirs, files in os.walk(directory):
+            for name in files:
+                if Path(name).suffix == ".csv" and args.target not in name and "excluded" not in name:
+                    print("Loading {}".format(name))
+                    train.append(pd.read_csv(os.path.join(root, name), sep="\t", header=0))
 
-    train = pd.concat(train, axis=0)
-    train.to_csv(f"{args.output_dir}/{target}_excluded_dataset.csv", sep="\t", index=False)
+        train = pd.concat(train, axis=0)
+        train.to_csv(f"{args.output_dir}/{target}_excluded_dataset.csv", sep="\t", index=False)
